LayoutTests/fast/canvas/webgl/drawingbuffer-test-expected.txt

This test ensures WebGL implementations correctly implement drawingbufferWidth/Height.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context exists
Checking drawingBufferWidth/drawingBufferHeight
PASS maxSize[0] > 0 is true
PASS maxSize[1] > 0 is true
PASS gl.drawingBufferWidth > 0 is true
PASS gl.drawingBufferHeight > 0 is true
PASS gl.drawingBufferWidth <= maxSize[0] is true
PASS gl.drawingBufferHeight <= maxSize[1] is true

PASS context exists
Checking drawingBufferWidth/drawingBufferHeight
PASS maxSize[0] > 0 is true
PASS maxSize[1] > 0 is true
PASS gl.drawingBufferWidth > 0 is true
PASS gl.drawingBufferHeight > 0 is true
PASS gl.drawingBufferWidth <= maxSize[0] is true
PASS gl.drawingBufferHeight <= maxSize[1] is true

PASS context exists
Checking drawingBufferWidth/drawingBufferHeight
PASS maxSize[0] > 0 is true
PASS maxSize[1] > 0 is true
PASS gl.drawingBufferWidth > 0 is true
PASS gl.drawingBufferHeight > 0 is true
PASS gl.drawingBufferWidth <= maxSize[0] is true
PASS gl.drawingBufferHeight <= maxSize[1] is true

PASS successfullyParsed is true

TEST COMPLETE
